backend/requirements.txt
Flask
flask-cors
python-dotenv
google-auth
